pages: migrate the main index and admin pages
